Fix issue #2 - getNeededHeading()
This should fix the getNeededHeading() function, however I haven't been able to compile and test it so please test before merging. It changes the domain of the angle from [-pi,pi] to [0,2pi] and then adjusts for the difference in the way we do angles on a Cartesian plane vs compass headings (north being up). No need to use the current heading at all, since we only want the destination heading, not the difference.